backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.1
python-multipart==0.0.9
pillow==10.2.0
opencv-python==4.9.0.80
scikit-image==0.22.0
wand==0.6.13
numpy==1.26.4
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.0
pytest==8.0.0
httpx==0.26.0
